# Backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Omega School API
  version: 1.0.0
  description: API pour gérer une école

servers:
  - url: http://localhost:8080
    description: Serveur local
  - url: https://api-spring-boot-uweb.onrender.com
    description: Serveur en production

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Ressource non trouvée"
        path:
          type: string
          example: "/students/1234"

    Sex:
      type: string
      enum: [MALE, FEMALE]

    Role:
      type: string
      enum: [ADMIN, TEACHER, STUDENT]

    UserRequestDto:
      type: object
      required: [firstName, lastName, email, birthDate, sex, address, phoneNumber, password, role]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        birthDate: { type: string, format: date }
        sex: { $ref: '#/components/schemas/Sex' }
        address: { type: string }
        phoneNumber: { type: string }
        password: { type: string }
        role: { $ref: '#/components/schemas/Role' }

    StudentRequestDto:
      allOf:
        - $ref: '#/components/schemas/UserRequestDto'
        - type: object
          required: [registrationNumber, level, group]
          properties:
            registrationNumber: { type: string }
            level: { type: string }
            group: { type: string }
  
    TeacherRequestDto:
      allOf:
        - $ref: '#/components/schemas/UserRequestDto'
        - type: object
          required: [matriculeNumber]
          properties:
            matriculeNumber: { type: string }
            bio: { type: string }

    User:
      allOf:
        - $ref: '#/components/schemas/UserRequestDto'
        - type: object
          properties:
            userId: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }

    
    Student:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            registrationNumber: { type: string }
            level:
              type: object
              properties:
                name: { type: string }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
            group:
              type: object
              properties:
                name: { type: string }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }

    Teacher:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            matriculeNumber: { type: string }
            bio: { type: string }

paths:
  /users:
    post:
      summary: Créer un nouvel utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRequestDto' }
      responses:
        '201': { description: Utilisateur créé, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { description: Requête invalide, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    get:
      summary: Récupérer tous les utilisateurs
      responses:
        '200': { description: Liste des utilisateurs, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } }
        '204': { description: Aucun utilisateur trouvé }

  /users/{id}:
    get:
      summary: Récupérer un utilisateur par ID
      parameters:
        - in: path
          name: id
          schema: { type: string, format: uuid }
          required: true
      responses:
        '200': { description: Utilisateur trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '404': { description: Utilisateur non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    put:
      summary: Mettre à jour un utilisateur
      parameters:
        - in: path
          name: id
          schema: { type: string, format: uuid }
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRequestDto' }
      responses:
        '200': { description: Utilisateur mis à jour, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { description: Données invalides, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Utilisateur non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    delete:
      summary: Supprimer un utilisateur
      parameters:
        - in: path
          name: id
          schema: { type: string, format: uuid }
          required: true
      responses:
        '204': { description: Utilisateur supprimé }
        '404': { description: Utilisateur non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /users/email/{email}:
    get:
      summary: Récupérer un utilisateur par email
      parameters:
        - in: path
          name: email
          schema: { type: string, format: email }
          required: true
      responses:
        '200': { description: Utilisateur trouvé, content: {application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '404': { description: Utilisateur non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /users/role/{role}:
    get:
      summary: Récupérer les utilisateurs par rôle
      parameters:
        - in: path
          name: role
          schema: { $ref: '#/components/schemas/Role' }
          required: true
      responses:
        '200': { description: Liste des utilisateurs, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } }
        '204': { description: Aucun utilisateur trouvé }

  /teachers:
    post:
      summary: Créer un nouvel enseignant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeacherRequestDto' }
      responses:
        '201': { description: Enseignant créé, content: { application/json: { schema: { $ref: '#/components/schemas/Teacher' } } } }
        '400': { description: Requête invalide, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    get:
      summary: Récupérer tous les enseignants
      responses:
        '200': { description: Liste des enseignants, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Teacher' } } } } }
        '204': { description: Aucun enseignant trouvé }

  /teachers/{userId}:
    get:
      et:
      summary: Récupérer un enseignant par ID
      parameters:
        - in: path
          name: userId
          schema: { type: string, format: uuid }
          required: true
      responses:
        '200': { description: Enseignant trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/Teacher' } } } }
        '404': { description: Enseignant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    put:
      summary: Mettre à jour un enseignant
      parameters:
        - in: path
          name: userId
          schema: { type: string, format: uuid}
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeacherRequestDto' }
      responses:
        '200': { description: Enseignant mis à jour, content: { application/json: { schema: { $ref: '#/components/schemas/Teacher' } } } }
        '404': { description: Enseignant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    delete:
      summary: Supprimer un enseignant
      parameters:
        - in: path
          name: userId
          schema: { type: string, format: uuid }
          required: true
      responses:
        '204': { description: Enseignant supprimé }
        '404': { description: Enseignant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /teachers/matricule/{matricule}:
    get:
      summary: Récupérer un enseignant par matricule
      parameters:
        - in: path
          name: matricule
          schema: { type: string }
          required: true
      responses:
        '200': { description: Enseignant trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/Teacher' } } } }
        '404': { description: Enseignant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /students:
    post:
      summary: Créer un nouvel étudiant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentRequestDto' }
      responses:
        '201': { description: Étudiant créé, content: { application/json: { schema: { $ref: '#/components/schemas/Student' } } } }
        '400': { description: Requête invalide, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    get:
      summary: Récupérer tous les étudiants
      responses:
        '200': { description: Liste des étudiants, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Student' } } } } }
        '204': { description: Aucun étudiant trouvé }

  /students/{userId}:
    get:
      summary: Récupérer un étudiant par ID
      parameters:
        - in: path
          name: userId
          schema: { type: string, format: uuid }
          required: true
      responses:
        '200': { description: Étudiant trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/Student' } } } }
        '404': { description: Étudiant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    put:
      summary: Mettre à jour un étudiant
      parameters:
        - in: path
          name: userId
          schema: { type: string, format: uuid }
          required: true
      requestBody:
        required: true
        content: {  application/json: { schema: { $ref: '#/components/schemas/StudentRequestDto'} } }
      responses:
        '200': { description: Étudiant mis à jour, content: { application/json: { schema: { $ref: '#/components/schemas/Student' } } } }
        '404': { description: Étudiant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    delete:
      summary: Supprimer un étudiant
      parameters:
        - in: path
          name: userId
          schema: { type: string, format: uuid }
          required: true
      responses:
        '204': { description: Étudiant supprimé }
        '404': { description: Étudiant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /students/registration/{regNumber}:
    get:
      summary: Récupérer un étudiant par numéro d'inscription
      parameters:
        - in: path
          name: regNumber
          schema: {  type: string }
          required: true
      responses:
        '200': { description: Étudiant trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/Student' } } } }
        '404': { description: Étudiant non trouvé, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        
  /students/level/{level}:
    get:
      summary: Récupérer les étudiants par niveau
      parameters:
        - in: path
          name: level
          schema: { type: string }
          required: true
      responses:
        '200': { description: Étudiants trouvés, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Student' } } } } }
        '404': { description: Aucun étudiant trouvé pour ce niveau, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /students/group/{group}:
    get:
      summary: Récupérer les étudiants par groupe
      parameters:
        - in: path
          name: group
          schema: { type: string }
          required: true
      responses:
        '200': { description: Étudiants trouvés, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Student' } } } } }
        '404': { description: Aucun étudiant trouvé pour ce groupe, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }